refactor: move actions to lib folder
